Creates new Topshelf.Dashboard assembly
Topshelf requires Spark view engine only for dashboard support.
Projects which doesn't want it shouldn't need to distribute Spark assemblies.
Also, this allow another dashboards implementations (Razor, anyone?)

* Remove NewtonSoft.Json (unused?) and Spark references from TopShelf assembly
* Change Topshelf.Host default namespace from "Topshelf" to "Topshelf.Host"
* Add Topshelf.Dashboard reference to Topshelf.Host
* Add Topshelf.Dashboard reference to Stuff sample project
* Update namespaces to match correct Assemblies/Folders
* Minor cosmetic fixes (mostly unused "using" directives)
