+++ src/Samples/Stuff/Stuff.csproj
@@ -69,6 +69,10 @@
     <Compile Include="TownCrier.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\Topshelf.Dashboard\Topshelf.Dashboard.csproj">
+      <Project>{67C7C55C-A602-4E95-B75F-D37194E9F676}</Project>
+      <Name>Topshelf.Dashboard</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\Topshelf\Topshelf.csproj">
       <Project>{1EEC6A0D-F213-4716-B5CC-33782BCE1908}</Project>
       <Name>Topshelf</Name>
