+++ src/Topshelf.Dashboard/Topshelf.Dashboard.csproj
@@ -1,0 +1,106 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProductVersion>8.0.30703</ProductVersion>
+    <SchemaVersion>2.0</SchemaVersion>
+    <ProjectGuid>{67C7C55C-A602-4E95-B75F-D37194E9F676}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>Topshelf.Dashboard</RootNamespace>
+    <AssemblyName>Topshelf.Dashboard</AssemblyName>
+    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
+    <FileAlignment>512</FileAlignment>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>..\bin\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="log4net, Version=1.2.10.0, Culture=neutral, PublicKeyToken=1b44e1d426115821, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\lib\log4net.dll</HintPath>
+    </Reference>
+    <Reference Include="Magnum, Version=1.1.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\lib\Magnum\Magnum.dll</HintPath>
+    </Reference>
+    <Reference Include="Spark">
+      <HintPath>..\..\lib\Spark\Spark.dll</HintPath>
+    </Reference>
+    <Reference Include="Stact, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\lib\Stact\Stact.dll</HintPath>
+    </Reference>
+    <Reference Include="Stact.ServerFramework, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\lib\Stact\Stact.ServerFramework.dll</HintPath>
+    </Reference>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="DashboardConnectionHandler.cs" />
+    <Compile Include="DashboardExtensions.cs" />
+    <Compile Include="DashboardView.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ResponseContextExtensions.cs" />
+    <Compile Include="SparkRender.cs" />
+    <Compile Include="StaticResourceHandler.cs" />
+    <Compile Include="TopshelfDashboard.cs" />
+    <Compile Include="TopshelfView.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="images\control_pause.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="images\control_play.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="images\control_stop.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="styles\ie.css" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="styles\print.css" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="styles\screen.css" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="views\dashboard.html" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Topshelf\Topshelf.csproj">
+      <Project>{1EEC6A0D-F213-4716-B5CC-33782BCE1908}</Project>
+      <Name>Topshelf</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
