+++ src/Topshelf.Dashboard/images/control_pause.png
