+++ src/Topshelf.Dashboard/images/control_play.png
