+++ src/Topshelf.Dashboard/images/control_stop.png
