+++ src/Topshelf.Dashboard/views/dashboard.html
@@ -1,0 +1,35 @@
+﻿<viewdata model="DashboardView" />
+<!doctype html>
+<html>
+	<head>
+		<title>Topshelf Dashboard</title>
+        <link rel="stylesheet" href="styles/screen.css" type="text/css" media="screen, projection">
+        <link rel="stylesheet" href="styles/print.css" type="text/css" media="print">	
+        <!--[if lt IE 8]><link rel="stylesheet" href="styles/ie.css" type="text/css" media="screen, projection"><![endif]-->
+	</head>
+	<body>
+        <div class="container">
+	        <table>
+                <thead>
+                    <tr>
+                        <th>Service Id (guid)</th>
+                        <th>Name</th>
+                        <th>State</th>
+                        <th>Actions</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <tr each="var i in Model.Statuses">
+                        <td>${i.Name}</td>
+                        <td>${i.ServiceType}</td>
+                        <td>${i.CurrentState}</td>
+                        <td>
+                            <a href="#"><img src="images/control_play.png" alt="Play" /></a> | 
+                            <a href="#"><img src="images/control_stop.png" alt="Stop" /></a>
+                        </td>
+                    </tr>
+                </tbody>
+            </table>
+        </div>
+	</body>
+</html>
