+++ src/Topshelf.Host/Program.cs
@@ -10,10 +10,11 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the 
 // specific language governing permissions and limitations under the License.
-namespace Topshelf
+namespace Topshelf.Host
 {
 	using System;
 	using System.IO;
+	using Dashboard;
 	using log4net;
 	using log4net.Config;
