+++ src/Topshelf.sln
@@ -18,6 +18,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		..\rakefile.rb = ..\rakefile.rb
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Topshelf.Dashboard", "Topshelf.Dashboard\Topshelf.Dashboard.csproj", "{67C7C55C-A602-4E95-B75F-D37194E9F676}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -80,6 +82,16 @@ Global
 		{6C0B10B8-259E-4A1B-8415-4343F0B77054}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{6C0B10B8-259E-4A1B-8415-4343F0B77054}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{6C0B10B8-259E-4A1B-8415-4343F0B77054}.Release|x86.ActiveCfg = Release|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Release|Any CPU.Build.0 = Release|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{67C7C55C-A602-4E95-B75F-D37194E9F676}.Release|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
