--- src/Topshelf/Dashboard/ResponseContextExtensions.cs
@@ -1,27 +1,0 @@
-﻿// Copyright 2007-2011 The Apache Software Foundation.
-//  
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not use 
-// this file except in compliance with the License. You may obtain a copy of the 
-// License at 
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0 
-// 
-// Unless required by applicable law or agreed to in writing, software distributed 
-// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
-// CONDITIONS OF ANY KIND, either express or implied. See the License for the 
-// specific language governing permissions and limitations under the License.
-namespace Topshelf.Dashboard
-{
-	using Stact.ServerFramework;
-
-
-	public static class ResponseContextExtensions
-	{
-		public static void RenderSparkView<TViewData>(this ResponseContext cxt, TViewData data, string template)
-		{
-			var render = new SparkRender();
-			string output = render.Render(template, data);
-			cxt.WriteHtml(output);
-		}
-	}
-}
